University of Washington Discrimination Analysis
To explore whether there is any evidence suggestive of discrimination by sex in the employment of the faculty at University of Washington. Salary data was obtained on all faculty members employed by the University during the 1995 academic year.  Provided an analysis of 1995 salaries with the primary goal of
determining whether or not gender discrimination exists with respect to pay.