Case Analysis: SUPPORT Hospital Cost
The dataset used, was collected in the Study to Understand Prognoses and Preferences for Outcomes and Risks of Treatments (SUPPORT). The SUPPORT dataset contains information on a random sample of 1000 patients admitted into one of five academic medical centers across the US. While the original goal of the SUPPORT study was to identify factors that correlate with poor outcomes following hospital admission, it has also been used to examine factors that affect hospital costs.